requirements.txt
azure-core==1.35.1
azure-functions==1.24.0
azure-identity==1.25.1
azure-keyvault-secrets==4.10.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==46.0.2
dnspython==2.8.0
dotenv==0.9.9
idna==3.10
isodate==0.7.2
MarkupSafe==3.0.3
msal==1.34.0
msal-extensions==1.3.1
numpy==2.3.3
pandas==2.3.3
pyarrow==21.0.0
pycparser==2.23
PyJWT==2.10.1
pylance==0.38.1
pymongo==4.15.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
six==1.17.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
playwright
openpyxl
xlrd
python-calamine
